// apps/web/components/command-palette/constants.d.ts
export declare const TIMEZONE_GROUPS: {
    Popular: {
        value: string;
        label: string;
    }[];
    Americas: {
        value: string;
        label: string;
    }[];
    "Europe & Africa": {
        value: string;
        label: string;
    }[];
    "Asia & Pacific": {
        value: string;
        label: string;
    }[];
};
export declare const ALL_TIMEZONES: ({
    value: string;
    label: string;
} | {
    value: string;
    label: string;
} | {
    value: string;
    label: string;
} | {
    value: string;
    label: string;
})[];
export declare const WORKING_DAYS: {
    value: number;
    label: string;
}[];
export type PaletteView = "main" | "appearance" | "time-region" | "timezone" | "notifications" | "calendar-defaults" | "account" | "security" | "passkeys" | "calendars" | "calendar-create" | "calendar-edit" | "subscriptions" | "events" | "event-editor";
//# sourceMappingURL=constants.d.ts.map